+++ renderer/scripts/types/window.d.ts
@@ -1,0 +1,9 @@
+import { ElectronAPI } from '../../src/preload';
+
+declare global {
+  interface Window {
+    electronAPI: ElectronAPI;
+  }
+}
+
+export {};